+++ unit_test/batched/Test_Batched_TeamLU.hpp
@@ -33,7 +33,7 @@ namespace Test {
       auto aa = Kokkos::subview(_a, k, Kokkos::ALL(), Kokkos::ALL());
 
       if (member.team_rank() == 0) {
-        for (int i=0;i<aa.dimension_0();++i)                                                                          
+        for (int i=0;i<static_cast<int>(aa.dimension_0());++i)                                                                          
           aa(i,i) += 10.0;  
       }
       member.team_barrier();
